# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse_null_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
...
			"Role": {
				"enum": [
					"ADMIN",
					"USER",
					"ORGANIZER",
					"PROBLEM_SETTER"
				],
				"type": "string"
			},
			"Problem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"title": {
						"type": "string"
					},
					"statement": {
						"type": "string"
					},
					"difficulty": {
						"type": "number",
						"format": "double"
					},
					"timeLimit": {
						"type": "number",
						"format": "double"
					},
					"memoryLimit": {
						"type": "number",
						"format": "double"
					},
					"inputFormat": {
						"type": "string"
					},
					"outputFormat": {
						"type": "string"
					},
					"solutionText": {
						"type": "string"
					},
					"owner": {
						"$ref": "#/components/schemas/User"
					},
					"submissions": {
						"items": {
							"$ref": "#/components/schemas/Submission"
						},
						"type": "array"
					},
					"testcases": {
						"items": {
							"$ref": "#/components/schemas/Testcase"
						},
						"type": "array"
					},
					"isPublished": {
						"type": "boolean"
					},
					"associatedContests": {
						"items": {
							"$ref": "#/components/schemas/ProblemInContest"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"title",
					"statement",
					"difficulty",
					"timeLimit",
					"memoryLimit",
					"inputFormat",
					"outputFormat",
					"solutionText",
					"owner",
					"submissions",
					"testcases",
					"isPublished",
					"associatedContests"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"handle": {
						"type": "string"
					},
					"avatar_image": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"isBan": {
						"type": "boolean"
					},
					"lastTimeChangeHandle": {
						"type": "number",
						"format": "double"
					},
					"lastTimeChangeImage": {
						"type": "number",
						"format": "double"
					},
					"country": {
						"type": "string"
					},
					"problems": {
						"items": {
							"$ref": "#/components/schemas/Problem"
						},
						"type": "array"
					},
					"submissions": {
						"items": {
							"$ref": "#/components/schemas/Submission"
						},
						"type": "array"
					},
					"organizedContests": {
						"items": {
							"$ref": "#/components/schemas/Contest"
						},
						"type": "array"
					},
					"participatedContest": {
						"items": {
							"$ref": "#/components/schemas/ContestParticipation"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"handle",
					"avatar_image",
					"bio",
					"email",
					"password",
					"role",
					"isBan",
					"lastTimeChangeHandle",
					"lastTimeChangeImage",
					"country",
					"problems",
					"submissions",
					"organizedContests",
					"participatedContest"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Contest": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"ruleText": {
						"type": "string"
					},
					"startTime": {
						"type": "number",
						"format": "double"
					},
					"endTime": {
						"type": "number",
						"format": "double"
					},
					"scoringRule": {
						"type": "string"
					},
					"isPlagiarismCheckEnabled": {
						"type": "boolean"
					},
					"isPublished": {
						"type": "boolean"
					},
					"organizer": {
						"$ref": "#/components/schemas/User"
					},
					"participations": {
						"items": {
							"$ref": "#/components/schemas/ContestParticipation"
						},
						"type": "array"
					},
					"problemsInContest": {
						"items": {
							"$ref": "#/components/schemas/ProblemInContest"
						},
						"type": "array"
					},
					"submissions": {
						"items": {
							"$ref": "#/components/schemas/Submission"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"title",
					"description",
					"ruleText",
					"startTime",
					"endTime",
					"scoringRule",
					"isPlagiarismCheckEnabled",
					"isPublished",
					"organizer",
					"participations",
					"problemsInContest",
					"submissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ContestParticipation": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"contestId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"contest": {
						"$ref": "#/components/schemas/Contest"
					},
					"score": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"userId",
					"contestId",
					"user",
					"contest",
					"score"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProblemInContest": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"problem": {
						"$ref": "#/components/schemas/Problem"
					},
					"contest": {
						"$ref": "#/components/schemas/Contest"
					},
					"score": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"problem",
					"contest",
					"score"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Submission": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"owner": {
						"$ref": "#/components/schemas/User"
					},
					"problem": {
						"$ref": "#/components/schemas/Problem"
					},
					"contest": {
						"$ref": "#/components/schemas/Contest"
					},
					"code": {
						"type": "string"
					},
					"language": {
						"$ref": "#/components/schemas/LANGUAGE"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/SubmissionResult"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"owner",
					"problem",
					"contest",
					"code",
					"language",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LANGUAGE": {
				"type": "string",
				"enum": [
					"CPP",
					"C",
					"JAVA",
					"PYTHON"
				]
			},
			"Testcase": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"problem": {
						"$ref": "#/components/schemas/Problem"
					},
					"input": {
						"type": "string"
					},
					"output": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"problem",
					"input",
					"output"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SubmissionResult": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"submission": {
						"$ref": "#/components/schemas/Submission"
					},
					"testcase": {
						"$ref": "#/components/schemas/Testcase"
					},
					"result": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"submission",
					"testcase",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"NewProblemRequest": {
				"properties": {
					"isPublished": {
						"type": "boolean"
					},
					"solutionText": {
						"type": "string"
					},
					"outputFormat": {
						"type": "string"
					},
					"inputFormat": {
						"type": "string"
					},
					"memoryLimit": {
						"type": "number",
						"format": "double"
					},
					"timeLimit": {
						"type": "number",
						"format": "double"
					},
					"difficulty": {
						"type": "number",
						"format": "double"
					},
					"statement": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"outputFormat",
					"inputFormat",
					"memoryLimit",
					"timeLimit",
					"statement",
					"title"
				],
				"type": "object"
			},
			"UpdateProblemRequest": {
				"properties": {
					"isPublished": {
						"type": "boolean"
					},
					"solutionText": {
						"type": "string"
					},
					"outputFormat": {
						"type": "string"
					},
					"inputFormat": {
						"type": "string"
					},
					"memoryLimit": {
						"type": "number",
						"format": "double"
					},
					"timeLimit": {
						"type": "number",
						"format": "double"
					},
					"difficulty": {
						"type": "number",
						"format": "double"
					},
					"statement": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"GetProblemResponse": {
				"$ref": "#/components/schemas/Problem"
			},
			"ApiResponse_GetProblemResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetProblemResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
			"GetAllProblemsResponse": {
				"items": {
					"$ref": "#/components/schemas/Problem"
				},
				"type": "array"
			},
			"ApiResponse_GetAllProblemsResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetAllProblemsResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
...
			"GetContestProblemResponse": {
				"items": {
					"$ref": "#/components/schemas/Problem"
				},
				"type": "array"
			},
			"ApiResponse_GetContestProblemResponse_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/GetContestProblemResponse"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object"
			},
			"DeleteProblemInContestRequest": {
				"properties": {
					"problemId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"problemId"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "web-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/problem": {
			"post": {
				"operationId": "CreateProblem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewProblemRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllProblems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetAllProblemsResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/problem/{id}": {
			"patch": {
				"operationId": "UpdateProblem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProblemRequest"
							}
						}
					}
				}
			}
		},
		"/problem/{id}/testcase": {
			"post": {
				"operationId": "AddTestcase",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"output": {
										"type": "string"
									},
									"input": {
										"type": "string"
									}
								},
								"required": [
									"output",
									"input"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/problem/{id}/testcases": {
			"post": {
				"operationId": "AddTestcases",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"testcases": {
										"items": {
											"properties": {
												"output": {
													"type": "string"
												},
												"input": {
													"type": "string"
												}
											},
											"required": [
												"output",
												"input"
											],
											"type": "object"
										},
										"type": "array"
									}
								},
								"required": [
									"testcases"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/problem/id/{id}": {
			"get": {
				"operationId": "GetProblemById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetProblemResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "SoftDeleteProblem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "header",
						"name": "x-access-token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/problem/title/{title}": {
			"get": {
				"operationId": "GetProblemByTitle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetProblemResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "title",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/problem/search": {
			"get": {
				"operationId": "SearchProblems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetAllProblemsResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Problem"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "searchKeyword",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "difficultyLow",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "difficultyHigh",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/contest/{id}/problem": {
			"get": {
				"operationId": "GetContestProblems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GetContestProblemResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteProblem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteProblemInContestRequest"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "EditProblem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"score": {
										"type": "number",
										"format": "double"
									},
									"problemId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"problemId"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"post": {
				"operationId": "AddProblem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Contest"
				],
				"security": [
					{
						"jwt": [
							"USER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"score": {
										"type": "number",
										"format": "double"
									},
									"problemId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"problemId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}